Moved HoloParams and get_phase to reconstr file
