+++ src/libertem_holo/base/reconstr.py
@@ -31,13 +31,19 @@ This includes both FFT-based and phase-shifting approaches.
 from __future__ import annotations
 
 from typing import Any, Literal
-
+import typing
+import time
+from sparseconverter import NUMPY, for_backend
 import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.colors import LogNorm
 from numpy.fft import fft2
+import logging
 
+from libertem_holo.base.filters import phase_unwrap
 from libertem_holo.base.mask import disk_aperture
+
+log = logging.getLogger(__name__)
 
 XPType = Any  # Union[Module("numpy"), Module("cupy")]
 
@@ -462,3 +468,52 @@ def display_fft_image(image, sb_position, slice_fft, mask=1, detail=True):
         ax1.set_title("Without Aperture Mask")
         ax2.imshow(np.abs(fft_with_aperture), vmax=0.01, cmap="gray")
         ax2.set_title("FFT with Aperture Mask")
+
+
+class HoloParams(typing.NamedTuple):
+    sb_size: tuple[float, float]
+    sb_position: tuple[float, float]
+    aperture: np.ndarray  # actually can be a cupy ndarray, too! hm...
+    orig_shape: tuple[int, int]
+    out_shape: tuple[int, int]
+    scale_factor: float  # by how much is the phase image scaled down in comparison to the hologram?
+
+    @property
+    def sb_position_raw(self) -> tuple[float, float]:
+        # the actual sb_position contains arrays, convert them to floats:
+        return tuple(
+            float(for_backend(c, NUMPY))
+            for c in self.sb_position
+        )
+
+    @property
+    def sb_position_int(self) -> tuple[int, int]:
+        return tuple(
+            int(c)
+            for c in self.sb_position_raw
+        )
+
+
+def get_phase(hologram, params: HoloParams, xp=np) -> np.ndarray:
+    t0 = time.perf_counter()
+
+    slice_fft = get_slice_fft(params.out_shape, hologram.shape)
+    phase_amp = reconstruct_frame(
+        hologram,
+        sb_pos=params.sb_position,
+        aperture=params.aperture,
+        slice_fft=slice_fft,
+        xp=xp
+    )
+
+    # phase_unwrap is numpy-only:
+    phase = for_backend(np.angle(phase_amp), NUMPY)
+
+    t1 = time.perf_counter()
+    
+    phase_unwrapped = phase_unwrap(phase)
+
+    t2 = time.perf_counter()
+
+    log.debug(f"time: total={t2-t0:.3f}s, reconstruction={t1-t0:.3f}s, unwrapping={t2-t1:.3f}s")
+    return phase_unwrapped
